+++ packages/zest-ui/src/menu/portal/MenuPortal.tsx
@@ -8,6 +8,7 @@ import { createChangeEventDetails } from '../../utils/createChangeEventDetails';
 import { REASONS } from '../../utils/reasons';
 import type { ZestPortalModalProps } from '../../types';
 import { useStoreState } from '../../store/ReactStore';
+import { PortalGestureRoot } from '../../internals/portal/PortalGestureRoot';
 
 /**
  * A portal element that moves the popup to the top of the app.
@@ -53,7 +54,7 @@ export function MenuPortal(props: MenuPortal.Props) {
           }
         }}
       >
-        {children}
+        <PortalGestureRoot>{children}</PortalGestureRoot>
       </Modal>
     </MenuPortalContext.Provider>
   );